feat: delete main app in place of web
