--- apps/main/.env
@@ -1,1 +1,0 @@
-DOCS_URL="https://wells-docs.vercel.app"
