--- apps/main/.env.development
@@ -1,2 +1,0 @@
-BLOG_URL="http://localhost:3002"
-DOCS_URL="http://localhost:3001"
