--- apps/main/.vscode/settings.json
@@ -1,4 +1,0 @@
-{
-  "typescript.tsdk": "../../node_modules/.pnpm/typescript@4.9.4/node_modules/typescript/lib",
-  "typescript.enablePromptUseWorkspaceTsdk": true
-}
