--- apps/main/CHANGELOG.md
@@ -1,91 +1,0 @@
-# main
-
-## 0.1.0
-
-### Minor Changes
-
-- 9389483: Tailwind updates
-
-## 0.0.12
-
-### Patch Changes
-
-- 1b6e96c: Updated package.json dependency patterns
-
-## 0.0.11
-
-### Patch Changes
-
-- 3df9de9: Minor cleanup to scripts
-- Updated dependencies [3df9de9]
-  - ui@0.0.5
-
-## 0.0.10
-
-### Patch Changes
-
-- Updated dependencies [f1ad37b]
-- Updated dependencies [0da389b]
-  - ui@0.0.4
-
-## 0.0.9
-
-### Patch Changes
-
-- Updated dependencies [a89b4db]
-  - ui@0.0.3
-
-## 0.0.8
-
-### Patch Changes
-
-- 5442d31: Upgrade to Next 13
-  - ui@0.0.2
-
-## 0.0.7
-
-### Patch Changes
-
-- 01be378: Basic font optimization and some minimal content while I work.
-
-## 0.0.6
-
-### Patch Changes
-
-- 481946e: Move files to `src` and introduce vanilla-extract for styles
-
-## 0.0.5
-
-### Patch Changes
-
-- 24bf238: Adds favicons to each app
-
-## 0.0.4
-
-### Patch Changes
-
-- a9cdde2: Adds docs to multizone host routing
-- 07e2269: Adds support in main next.config.js for routing docs
-
-## 0.0.3
-
-### Patch Changes
-
-- af3bf63: Barebones multizonal approach using main as the host app and a nextra blog as the blog.
-- Updated dependencies [af3bf63]
-  - ui@0.0.2
-
-## 0.0.2
-
-### Patch Changes
-
-- da19d98: Update linting config to corrected configurations across next and react apps
-  - ui@0.0.1
-
-## 0.0.1
-
-### Patch Changes
-
-- 5a89b05: Init all packages and apps
-- Updated dependencies [5a89b05]
-  - ui@0.0.1
