--- apps/main/README.md
@@ -1,4 +1,0 @@
-# Wells Main
-
-Host multizonal app for my monorepo.
-
