--- apps/main/next.config.js
@@ -1,32 +1,0 @@
-const { BLOG_URL, DOCS_URL } = process.env;
-
-/** @type {import('next').NextConfig} */
-module.exports = {
-  reactStrictMode: true,
-  experimental: {
-    appDir: true,
-  },
-  async rewrites() {
-    return [
-      /**
-       * Rewrites for Multi Zones
-       */
-      {
-        source: "/blog",
-        destination: `${BLOG_URL}/blog`,
-      },
-      {
-        source: "/blog/:path*",
-        destination: `${BLOG_URL}/blog/:path*`,
-      },
-      {
-        source: "/docs",
-        destination: `${DOCS_URL}/docs`,
-      },
-      {
-        source: "/docs/:path*",
-        destination: `${DOCS_URL}/docs/:path*`,
-      },
-    ];
-  },
-};
