--- apps/main/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "main",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "clean": "rimraf .next && rimraf node_modules",
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint",
-    "lint:fix": "next lint --fix"
-  },
-  "dependencies": {
-    "next": "workspace:*",
-    "react": "workspace:*",
-    "react-dom": "workspace:*"
-  },
-  "devDependencies": {
-    "@types/node": "^18.11.18",
-    "@types/react": "workspace:*",
-    "autoprefixer": "^10.4.13",
-    "eslint": "workspace:*",
-    "eslint-config-custom": "workspace:*",
-    "postcss": "^8.4.21",
-    "tailwindcss": "^3.2.7",
-    "tsconfig": "workspace:*",
-    "typescript": "^4.9.4"
-  }
-}
