--- apps/main/postcss.config.js
@@ -1,6 +1,0 @@
-module.exports = {
-  plugins: {
-    tailwindcss: { config: "./tailwind.config.js" },
-    autoprefixer: {},
-  },
-};
