--- apps/main/src/app/layout.tsx
@@ -1,26 +1,0 @@
-import "../styles/globals.css";
-import { Inter } from "next/font/google";
-
-const inter = Inter({
-  subsets: ["latin"],
-  display: "swap",
-});
-
-export const metadata = {
-  title: "Wells",
-  description: "Wellscome to my website.",
-};
-
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <html lang="en" className={inter.className}>
-      <body className="bg-white dark:bg-zinc-900 text-zinc-900 dark:text-zinc-100">
-        <main>{children}</main>
-      </body>
-    </html>
-  );
-}
