+++ apps/web/public/favicon/about.txt
