+++ apps/web/public/favicon/android-chrome-512x512.png
