+++ apps/web/public/favicon/favicon-32x32.png
