+++ apps/web/public/favicon/favicon.ico
