+++ apps/web/public/favicon/site.webmanifest
