+++ apps/web/src/app/layout.tsx
@@ -1,18 +1,26 @@
-import './globals.css'
+import "../styles/globals.css";
+import { Inter } from "next/font/google";
+
+const inter = Inter({
+  subsets: ["latin"],
+  display: "swap",
+});
 
 export const metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
-}
+  title: "Wells",
+  description: "Wellscome to my website.",
+};
 
 export default function RootLayout({
   children,
 }: {
-  children: React.ReactNode
+  children: React.ReactNode;
 }) {
   return (
-    <html lang="en">
-      <body>{children}</body>
+    <html lang="en" className={inter.className}>
+      <body className="bg-white dark:bg-zinc-900 text-zinc-900 dark:text-zinc-100">
+        <main>{children}</main>
+      </body>
     </html>
-  )
+  );
 }
